fixes #9199
git-svn-id: https://svn.code.sf.net/p/specify/code/trunk/Specify@10966 ce9bf4bf-376f-1848-a496-a1c9fe3b1a39

+++ src/edu/ku/brc/specify/dbsupport/AccessionAutoNumberAlphaNum.java
@@ -111,7 +111,7 @@ public class AccessionAutoNumberAlphaNum extends AutoNumberGeneric
         Integer yearVal = null;
         if (yearPos != null && StringUtils.isNotEmpty(value) && value.length() >= yearPos.second)
         {
-            yearVal = extractIntegerValue(yearPos, value);
+            yearVal = extractIntegerValue(yearPos, value).intValue();
         }
 
         StringBuilder sb = new StringBuilder("SELECT a.accessionNumber FROM Accession a Join a.division dv Join dv.numberingSchemes ans WHERE ans.id = ");
